Extract FacetResultPresenter
Most of the functionality in UnifiedSearchPresenter is actually
concerned with presenting the facets. This commit extracts the
facet-related code into its own class.
